+++ packages-next/fields-document/src/DocumentEditor/relationship.tsx
@@ -5,7 +5,7 @@ import { ReactEditor, RenderElementProps } from 'slate-react';
 import { Transforms, Editor } from 'slate';
 
 import { jsx } from '@keystone-ui/core';
-import { useKeystone } from '@keystone-next/admin-ui/context';
+import { useKeystone } from '@keystone-next/keystone/admin-ui/context';
 import { RelationshipSelect } from '@keystone-next/fields/types/relationship/views/RelationshipSelect';
 
 import { ToolbarButton } from './primitives';