+++ packages-next/fields/src/types/relationship/views/cards/useItemState.tsx
@@ -1,6 +1,6 @@
 import { FieldMeta, ListMeta } from '@keystone-next/types';
 import { useCallback, useMemo, useState } from 'react';
-import { gql, useQuery } from '@keystone-next/admin-ui/apollo';
+import { gql, useQuery } from '@keystone-next/keystone/admin-ui/apollo';
 import { DataGetter, makeDataGetter } from '@keystone-next/admin-ui-utils';
 import { controller } from '../index';