+++ packages-next/keystone/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/App/index.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { Core } from '@keystone-ui/core';
+import { AppProps } from 'next/app';
+import { DocumentNode } from 'graphql';
+import { AdminConfig, FieldViews } from '@keystone-next/types';
+import { ErrorBoundary } from '../../../../admin-ui/components';
+import { KeystoneProvider } from '../../../../admin-ui/context';
+
+type AppConfig = {
+  adminConfig: AdminConfig;
+  adminMetaHash: string;
+  fieldViews: FieldViews;
+  lazyMetadataQuery: DocumentNode;
+};
+
+export const getApp =
+  (props: AppConfig) =>
+  ({ Component, pageProps }: AppProps) => {
+    return (
+      <Core>
+        <KeystoneProvider {...props}>
+          <ErrorBoundary>
+            <Component {...pageProps} />
+          </ErrorBoundary>
+        </KeystoneProvider>
+      </Core>
+    );
+  };