+++ packages-next/keystone/src/session/index.ts
@@ -188,7 +188,8 @@ export function storedSessions({
     let isConnected = false;
     return {
       async get({ req, createContext }) {
-        let sessionId = await get({ req, createContext });
+        const data = (await get({ req, createContext })) as { sessionId: string } | undefined;
+        const sessionId = data?.sessionId;
         if (typeof sessionId === 'string') {
           if (!isConnected) {
             await store.connect?.();
@@ -207,7 +208,8 @@ export function storedSessions({
         return start?.({ res, data: { sessionId }, createContext }) || '';
       },
       async end({ req, res, createContext }) {
-        let sessionId = await get({ req, createContext });
+        const data = (await get({ req, createContext })) as { sessionId: string } | undefined;
+        const sessionId = data?.sessionId;
         if (typeof sessionId === 'string') {
           if (!isConnected) {
             await store.connect?.();