+++ tests/api-tests/access-control/authed.test.ts
@@ -45,13 +45,13 @@ const expectNamedArray = <T extends { id: IdType }, N extends string>(
 
 multiAdapterRunners().map(({ before, after, provider }) =>
   describe(`Provider: ${provider}`, () => {
-    let keystone: any,
+    let disconnect: any,
       items: Record<string, { id: IdType; name: string }[]>,
       user: { id: IdType; name: string; yesRead: string; noRead: string },
       context: KeystoneContext;
     beforeAll(async () => {
       const _before = await before(setupKeystone);
-      keystone = _before.keystone;
+      disconnect = _before.disconnect;
       context = _before.context;
 
       // ensure every list has at least some data
@@ -78,7 +78,7 @@ multiAdapterRunners().map(({ before, after, provider }) =>
       })) as { id: IdType; name: string; yesRead: string; noRead: string };
     });
     afterAll(async () => {
-      await after(keystone);
+      await after(disconnect);
     });
 
     describe('create', () => {